+++ src/roslibpy/comm/comm.py
@@ -3,11 +3,7 @@ from __future__ import print_function
 import json
 import logging
 
-from autobahn.twisted.websocket import WebSocketClientFactory, WebSocketClientProtocol
-from twisted.internet.protocol import ReconnectingClientFactory
-
-from . import Message, ServiceResponse
-from .event_emitter import EventEmitterMixin
+from .. import Message, ServiceResponse
 
 LOGGER = logging.getLogger('roslibpy')
 
@@ -17,7 +13,7 @@ class RosBridgeException(Exception):
     pass
 
 
-class RosBridgeProtocol(WebSocketClientProtocol):
+class RosBridgeProtocol(object):
     """Implements the websocket client protocol to encode/decode JSON ROS Bridge messages."""
 
     def __init__(self, *args, **kwargs):
@@ -31,6 +27,15 @@ class RosBridgeProtocol(WebSocketClientProtocol):
         }
         # TODO: add handlers for op: status
 
+    def on_message(self, payload):
+        message = Message(json.loads(payload.decode('utf8')))
+        handler = self._message_handlers.get(message['op'], None)
+        if not handler:
+            raise RosBridgeException(
+                'No handler registered for operation "%s"' % message['op'])
+
+        handler(message)
+
     def send_ros_message(self, message):
         """Encode and serialize ROS Bridge protocol message.
 
@@ -39,9 +44,9 @@ class RosBridgeProtocol(WebSocketClientProtocol):
         """
         try:
             json_message = json.dumps(dict(message)).encode('utf8')
-            LOGGER.debug('Sending ROS message: %s', json_message)
+            LOGGER.debug('Sending ROS message|<pre>%s</pre>', json_message)
 
-            self.sendMessage(json_message)
+            self.send_message(json_message)
         except Exception as exception:
             # TODO: Check if it makes sense to raise exception again here
             # Since this is wrapped in many layers of indirection
@@ -74,26 +79,7 @@ class RosBridgeProtocol(WebSocketClientProtocol):
         json_message = json.dumps(dict(message)).encode('utf8')
         LOGGER.debug('Sending ROS service request: %s', json_message)
 
-        self.sendMessage(json_message)
-
-    def onConnect(self, response):
-        LOGGER.debug('Server connected: %s', response.peer)
-
-    def onOpen(self):
-        LOGGER.info('Connection to ROS MASTER ready.')
-        self.factory.ready(self)
-
-    def onMessage(self, payload, isBinary):
-        if isBinary:
-            raise NotImplementedError('Add support for binary messages')
-
-        message = Message(json.loads(payload.decode('utf8')))
-        handler = self._message_handlers.get(message['op'], None)
-        if not handler:
-            raise RosBridgeException(
-                'No handler registered for operation "%s"' % message['op'])
-
-        handler(message)
+        self.send_message(json_message)
 
     def _handle_publish(self, message):
         self.factory.emit(message['topic'], message['msg'])
@@ -122,41 +108,3 @@ class RosBridgeProtocol(WebSocketClientProtocol):
                 'Expected service name missing in service request')
 
         self.factory.emit(message['service'], message)
-
-    def onClose(self, wasClean, code, reason):
-        LOGGER.info('WebSocket connection closed: %s', reason)
-
-
-class RosBridgeClientFactory(EventEmitterMixin, ReconnectingClientFactory, WebSocketClientFactory):
-    """Factory to construct instance of the ROS Bridge protocol."""
-    protocol = RosBridgeProtocol
-
-    def __init__(self, *args, **kwargs):
-        super(RosBridgeClientFactory, self).__init__(*args, **kwargs)
-        self._proto = None
-        self.setProtocolOptions(closeHandshakeTimeout=5)
-
-    def on_ready(self, callback):
-        if self._proto:
-            callback(self._proto)
-        else:
-            self.once('ready', callback)
-
-    def ready(self, proto):
-        self._proto = proto
-        self.emit('ready', proto)
-
-    def startedConnecting(self, connector):
-        LOGGER.debug('Started to connect...')
-
-    def clientConnectionLost(self, connector, reason):
-        LOGGER.debug('Lost connection. Reason: %s', reason)
-        self.emit('close', self._proto)
-        ReconnectingClientFactory.clientConnectionLost(self, connector, reason)
-        self._proto = None
-
-    def clientConnectionFailed(self, connector, reason):
-        LOGGER.debug('Connection failed. Reason: %s', reason)
-        ReconnectingClientFactory.clientConnectionFailed(
-            self, connector, reason)
-        self._proto = None