+++ tests/test_param.py
@@ -31,8 +31,18 @@ def run_param_manipulation():
         ros_client.terminate()
 
     ros_client.call_later(1, verify)
-    ros_client.run_event_loop()
+    ros_client.run_forever()
 
 
 def test_param_manipulation():
     helpers.run_as_process(run_param_manipulation)
+
+
+if __name__ == '__main__':
+    import logging
+
+    logging.basicConfig(
+        level=logging.DEBUG, format='[%(thread)03d] %(asctime)-15s [%(levelname)s] %(message)s')
+    LOGGER = logging.getLogger('test')
+
+    run_param_manipulation()