+++ tests/test_rosapi.py
@@ -1,5 +1,6 @@
+import time
+
 import helpers
-import pytest
 
 from roslibpy import Ros
 
@@ -12,14 +13,16 @@ def run_rosapi_topics(*args, **kwargs):
     ros_client = Ros(*args, **kwargs)
 
     def callback(topic_list):
+        print(topic_list)
         assert('/rosout' in topic_list['topics'])
+        time.sleep(1)
         ros_client.terminate()
 
     def get_topics():
         ros_client.get_topics(callback)
 
     ros_client.on_ready(get_topics)
-    ros_client.run_event_loop()
+    ros_client.run_forever()
 
 
 def test_rosapi_topics():
@@ -31,8 +34,18 @@ def test_rosapi_topics_url():
 
 
 def test_connection_url_checks():
+    import pytest
     with pytest.raises(Exception):
         helpers.run_as_process(Ros, host)
 
     with pytest.raises(Exception):
         helpers.run_as_process(Ros, u'http://%s:%d' % (host, port))
+
+
+if __name__ == '__main__':
+    import logging
+
+    logging.basicConfig(level=logging.DEBUG, format='[%(thread)03d] %(asctime)-15s [%(levelname)s] %(message)s')
+    LOGGER = logging.getLogger('test')
+
+    run_rosapi_topics(host, port)