+++ tests/test_topic.py
@@ -1,5 +1,7 @@
 from __future__ import print_function
 
+import time
+
 import helpers
 
 from roslibpy import Message, Ros, Topic
@@ -13,16 +15,18 @@ def run_topic_pubsub():
 
     def receive_message(message):
         context['counter'] += 1
-        assert(message['data'] == 'test')
+        assert message['data'] == 'hello world', 'Unexpected message content'
 
         if context['counter'] == 3:
+            time.sleep(1)
             ros_client.terminate()
 
     def start_sending():
-        message = Message({'data': 'test'})
-        publisher.publish(message)
-        publisher.publish(message)
-        publisher.publish(message)
+        while True:
+            if not ros_client.is_connected:
+                break
+            publisher.publish(Message({'data': 'hello world'}))
+            time.sleep(0.1)
         publisher.unadvertise()
 
     def start_receiving():
@@ -30,8 +34,19 @@ def run_topic_pubsub():
 
     ros_client.on_ready(start_receiving, run_in_thread=True)
     ros_client.on_ready(start_sending, run_in_thread=True)
-    ros_client.run_event_loop()
+    ros_client.run_forever()
+
+    assert context['counter'] >= 3, 'Expected at least 3 messages but got ' + str(context['counter'])
 
 
 def test_topic_pubsub():
     helpers.run_as_process(run_topic_pubsub)
+
+
+if __name__ == '__main__':
+    import logging
+
+    logging.basicConfig(level=logging.DEBUG, format='[%(thread)03d] %(asctime)-15s [%(levelname)s] %(message)s')
+    LOGGER = logging.getLogger('test')
+
+    run_topic_pubsub()